Remove username from SSH URL in tests
For the tests were we test prompting for the username we should make
sure not to include the username in the URL or as the username_ptr as
this will provide a default value. Will only be a problem for us when
prompting for a username is the default setting.